+++ src/app/common/ingest-via-file-upload.types.ts
@@ -1,0 +1,18 @@
+import { HttpHeaders } from "@angular/common/http";
+
+export interface UploadPrepareResponse {
+    uploadToken: string;
+    uploadUrl: string;
+    method: UploadAvailableMethod;
+    useMultipart: boolean;
+    headers: [string, string][];
+    fields: [string, string][];
+}
+
+export interface UploadPerareData {
+    uploadPrepareResponse: UploadPrepareResponse;
+    bodyObject: FormData | File;
+    uploadHeaders: HttpHeaders;
+}
+
+export type UploadAvailableMethod = "POST" | "PUT";
