+++ src/app/dataset-view/additional-components/overview-component/components/file-from-url-modal/file-from-url-modal.component.html
@@ -21,7 +21,7 @@
     <button
         type="button"
         class="btn btn-success"
-        (click)="saveFileUrl(); activeModal.close()"
+        (click)="activeModal.close()"
         data-test-id="save-file-url"
         [disabled]="fileUrlForm.invalid"
     >
