+++ src/app/dataset-view/additional-components/overview-component/overview.component.scss
@@ -26,17 +26,6 @@
         cursor: pointer;
     }
 
-    // .file-uploader {
-    //     opacity: 0;
-    //     padding: 4.5rem;
-    //     width: inherit;
-    //     height: 100%;
-
-    //     &:hover {
-    //         cursor: pointer;
-    //     }
-    // }
-
     .upload-file-icon {
         z-index: 10;
     }
